+++ packages/client/src/hooks/reduxTsHook.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux'
-import { AppDispath, RootState } from './store'
+import { AppDispath, RootState } from '../store'
 
 export const useAppDispatch = () => useDispatch<AppDispath>()
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
